move GithubClient to module
